+++ python/test/state_representation/space/test_spatial_state.py
@@ -27,11 +27,10 @@ class TestState(unittest.TestCase):
         self.assertEqual(state1.get_reference_frame(), "world")
         self.assertTrue(state1.is_empty())
 
-        state2 = SpatialState("test", "robot", False)
+        state2 = SpatialState("test", "robot")
         self.assertEqual(state2.get_type(), StateType.SPATIAL_STATE)
         self.assertEqual(state2.get_name(), "test")
         self.assertEqual(state2.get_reference_frame(), "robot")
-        self.assertFalse(state2.is_empty())
 
         state3 = SpatialState(state2)
         self.assertEqual(state2.get_type(), state3.get_type())
@@ -40,7 +39,7 @@ class TestState(unittest.TestCase):
         self.assertEqual(state2.is_empty(), state3.is_empty())
 
     def test_copy(self):
-        state = SpatialState("test", "robot", False)
+        state = SpatialState("test", "robot")
         for state_copy in [copy.copy(state), copy.deepcopy(state)]:
             self.assertEqual(state.get_type(), state_copy.get_type())
             self.assertEqual(state.get_name(), state_copy.get_name())
@@ -48,7 +47,7 @@ class TestState(unittest.TestCase):
             self.assertEqual(state.is_empty(), state_copy.is_empty())
 
     def test_compatibility(self):
-        state1 = SpatialState("test", "robot", True)
+        state1 = SpatialState("test", "robot")
         state2 = SpatialState("test")
         self.assertFalse(state1.is_compatible(state2))
         state2.set_reference_frame("robot")